Partition project into smaller modules.
